+++ ChangeLog.rst
@@ -8,6 +8,8 @@ dev
   (communicator should be passed via constructor)
 - Remove ``Function::operator[+-*/]`` to prevent memory corruption problems
   (does not affect Python interface)
+- Change ``PETScSNESSolver`` constructor to accept both MPI communicator and
+  type
 
 2016.2.0 [2016-11-30]
 ---------------------